refactor: вернуть InMemoryFilmStorage и переименовать интерфейсы
