+++ lombok.config
@@ -2,3 +2,4 @@ config.stopBubbling = true
 lombok.anyconstructor.addconstructorproperties = false
 lombok.addLombokGeneratedAnnotation = true
 lombok.addSuppressWarnings = false
+lombok.copyableAnnotations += org.springframework.beans.factory.annotation.Qualifier
