+++ src/main/java/ru/yandex/practicum/filmorate/service/mpa/MpaServiceImpl.java
@@ -4,7 +4,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import ru.yandex.practicum.filmorate.dto.MpaDTO;
 import ru.yandex.practicum.filmorate.exception.NotFoundException;
-import ru.yandex.practicum.filmorate.storage.mpa.MpaDbStorage;
+import ru.yandex.practicum.filmorate.storage.mpa.MpaStorage;
 
 import java.util.Collection;
 import java.util.stream.Collectors;
@@ -13,7 +13,7 @@ import java.util.stream.Collectors;
 @Service
 @RequiredArgsConstructor
 public class MpaServiceImpl implements MpaService {
-    private final MpaDbStorage mpaStorage;
+    private final MpaStorage mpaStorage;
 
     public MpaDTO getById(Integer mpaId) {
         return mpaStorage.findById(mpaId)
