+++ src/main/java/ru/yandex/practicum/filmorate/storage/like/LikeDbStorage.java
@@ -1,14 +1,43 @@
 package ru.yandex.practicum.filmorate.storage.like;
 
+import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Pageable;
+import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.stereotype.Repository;
 import ru.yandex.practicum.filmorate.model.like.Like;
 
 import java.util.List;
 
-public interface LikeDbStorage {
-    void save(Like like);
+@Repository
+@RequiredArgsConstructor
+public class LikeDbStorage implements LikeStorage {
+    private final JdbcTemplate jdbcTemplate;
 
-    void delete(Like like);
+    @Override
+    public void delete(Like like) {
+        String sql = "DELETE FROM likes " +
+                "WHERE film_id = ? " +
+                "AND user_id = ?";
 
-    List<Long> findTopFilmIdByCountLikes(Pageable pageable);
+        jdbcTemplate.update(sql, like.getFilmId(), like.getUserId());
+    }
+
+    @Override
+    public void save(Like like) {
+        String sql = "INSERT INTO likes (film_id, user_id) VALUES (?, ?)";
+
+        jdbcTemplate.update(sql, like.getFilmId(), like.getUserId());
+    }
+
+    @Override
+    public List<Long> findTopFilmIdByCountLikes(Pageable pageable) {
+        String sql = "SELECT f.id " +
+                "FROM films f " +
+                "LEFT JOIN likes l ON f.id = l.film_id " +
+                "GROUP BY f.id " +
+                "ORDER BY COUNT(l.user_id) DESC " +
+                "LIMIT ?";
+
+        return jdbcTemplate.query(sql, (rs, rowNum) -> rs.getLong("id"), pageable.getPageSize());
+    }
 }
