+++ src/main/java/ru/yandex/practicum/filmorate/storage/mpa/MpaStorage.java
@@ -1,0 +1,12 @@
+package ru.yandex.practicum.filmorate.storage.mpa;
+
+import ru.yandex.practicum.filmorate.model.film.Mpa;
+
+import java.util.List;
+import java.util.Optional;
+
+public interface MpaStorage {
+    Optional<Mpa> findById(int mpaId);
+
+    List<Mpa> findAll();
+}
