+++ src/test/java/ru/yandex/practicum/filmorate/service/like/LikeServiceImplTest.java
@@ -9,9 +9,9 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.data.domain.PageRequest;
 import ru.yandex.practicum.filmorate.exception.NotFoundException;
 import ru.yandex.practicum.filmorate.model.like.Like;
-import ru.yandex.practicum.filmorate.storage.film.FilmDbStorage;
-import ru.yandex.practicum.filmorate.storage.like.LikeDbStorage;
-import ru.yandex.practicum.filmorate.storage.user.UserDbStorage;
+import ru.yandex.practicum.filmorate.storage.film.FilmStorage;
+import ru.yandex.practicum.filmorate.storage.like.LikeStorage;
+import ru.yandex.practicum.filmorate.storage.user.UserStorage;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
@@ -20,11 +20,11 @@ import static org.mockito.BDDMockito.given;
 @ExtendWith(MockitoExtension.class)
 class LikeServiceImplTest {
     @Mock
-    private LikeDbStorage likeStorage;
+    private LikeStorage likeStorage;
     @Mock
-    private FilmDbStorage filmStorage;
+    private FilmStorage filmStorage;
     @Mock
-    private UserDbStorage userStorage;
+    private UserStorage userStorage;
     private LikeService likeService;
 
     @BeforeEach
