+++ src/test/java/ru/yandex/practicum/filmorate/service/user/UserServiceImplTest.java
@@ -10,7 +10,7 @@ import ru.yandex.practicum.filmorate.dto.UserDTO;
 import ru.yandex.practicum.filmorate.exception.NotFoundException;
 import ru.yandex.practicum.filmorate.model.user.User;
 import ru.yandex.practicum.filmorate.service.friendship.FriendshipService;
-import ru.yandex.practicum.filmorate.storage.user.UserDbStorage;
+import ru.yandex.practicum.filmorate.storage.user.UserStorage;
 
 import java.time.LocalDate;
 import java.util.Collection;
@@ -26,7 +26,7 @@ import static org.mockito.BDDMockito.given;
 class UserServiceImplTest {
     private UserService userService;
     @Mock
-    private UserDbStorage userStorage;
+    private UserStorage userStorage;
     @Mock
     private FriendshipService friendshipService;
